MtebEval/requirements.txt
datasets
huggingface_hub
mteb
pyyaml